+++ entity/module.go
@@ -7,7 +7,7 @@ type Module struct {
 	tableName          string    `pg:"?db_schema.modules" json:"-"`
 	Id                 string    `json:"id"`
 	Name               string    `json:"name" valid:"required~Required"`
-	CreatedAt          time.Time `json:"createdAt" sql:",null"`
+	CreatedAt          time.Time `json:"createdAt" pg:",null"`
 	LastConnectedAt    time.Time `json:"lastConnectedAt"`
 	LastDisconnectedAt time.Time `json:"lastDisconnectedAt"`
 }